design fixes, full path for images
