+++ local-development/devtool/adapters/foreground.py
@@ -3,23 +3,13 @@
 from __future__ import annotations
 
 import os
-import shutil
 import subprocess
 import sys
 from pathlib import Path
 
-from devtool.domain.models import Service, ServiceType
+from devtool.domain.models import Service, ServiceType, WindowLayout
 from devtool.ports.session_manager import SessionManager
-
-
-def _resolve_bash() -> str:
-    """Find bash on PATH instead of assuming /bin/bash."""
-    path = shutil.which("bash")
-    if not path:
-        raise RuntimeError(
-            "bash not found on PATH. Install bash or set SHELL to a compatible shell."
-        )
-    return path
+from devtool.services.shell_utils import resolve_bash
 
 
 class ForegroundSessionManager(SessionManager):
@@ -47,7 +37,7 @@ class ForegroundSessionManager(SessionManager):
 
         cmd = commands[svc.name]
         shell_cmd = f"{cmd} 2>&1 | tee {log_path}"
-        bash = _resolve_bash()
+        bash = resolve_bash()
         os.execvp(bash, [bash, "-c", shell_cmd])
 
     def attach(self, session_name: str) -> None:
@@ -64,3 +54,6 @@ class ForegroundSessionManager(SessionManager):
 
     def pane_contents(self, session_name: str) -> dict[str, str]:
         return {}
+
+    def restart_service(self, session_name: str, service_name: str) -> bool:
+        return False
