+++ local-development/devtool/ports/process_manager.py
@@ -1,31 +1,24 @@
-"""Port: process manager for launching/killing processes."""
+"""Port: process and port management."""
 
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
-from pathlib import Path
+
+from devtool.domain.models import PortOccupant
 
 
 class ProcessManager(ABC):
 
     @abstractmethod
-    def kill_port(self, port: int) -> None:
-        """Kill any process listening on the given port."""
+    def find_port_occupants(self, port: int) -> list[PortOccupant]:
+        """Return PIDs and process names listening on *port*."""
 
     @abstractmethod
     def is_port_in_use(self, port: int) -> bool:
-        """Return True if something is listening on the port."""
+        """Return True if any process is bound to *port*."""
 
     @abstractmethod
-    def run_command(
-        self,
-        cmd: list[str],
-        cwd: Path,
-        *,
-        log_file: Path | None = None,
+    def kill_processes(
+        self, pids: list[int], *, graceful_timeout: float = 0.5,
     ) -> None:
-        """Run a command synchronously.
-
-        If *log_file* is provided, stdout is appended there and only stderr
-        is shown on the terminal.  Raises on non-zero exit.
-        """
+        """Send SIGTERM, wait *graceful_timeout*, then SIGKILL survivors."""
