+++ local-development/devtool/services/constants.py
@@ -1,0 +1,3 @@
+"""Application-layer constants shared across services."""
+
+SESSION_NAME: str = "unifai-dev"
