+++ local-development/devtool/services/recovery.py
@@ -2,8 +2,6 @@
 
 from __future__ import annotations
 
-import os
-import subprocess
 from pathlib import Path
 
 from devtool.domain.models import ContainerStatus, ServiceType
@@ -12,6 +10,7 @@ from devtool.ports.container_runtime import ContainerRuntime
 from devtool.ports.session_manager import SessionManager
 from devtool.ports.venv_manager import VenvManager
 from devtool.services import health_checker
+from devtool.services.constants import SESSION_NAME
 
 
 class Recovery:
@@ -47,12 +46,11 @@ class Recovery:
         if restarted_infra:
             print(f"  Restarted infra: {', '.join(restarted_infra)}\n")
 
-        # Send restart signal via tmux if session is running
-        from devtool.services.orchestrator import SESSION_NAME
-
         if self._session.is_running(SESSION_NAME):
-            self._send_restart_to_pane(SESSION_NAME, svc.name)
-            print(f"  ✔ Sent restart signal to {svc.name}")
+            if self._session.restart_service(SESSION_NAME, svc.name):
+                print(f"  ✔ Sent restart signal to {svc.name}")
+            else:
+                print(f"  ⚠ Could not find pane for {svc.name}")
         else:
             print(f"  ⚠ No active session — start services first.")
 
@@ -71,7 +69,6 @@ class Recovery:
             print("  ✔ All services are healthy.")
             return
 
-        # Restart in dependency order: infra first, then primaries, then workers
         primary_failed = [
             n for n in failed
             if self._registry.get_service(n).is_primary
@@ -83,37 +80,3 @@ class Recovery:
 
         for name in primary_failed + worker_failed:
             self.restart_service(name, root)
-
-    @staticmethod
-    def _send_restart_to_pane(session_name: str, service_name: str) -> None:
-        """Send Ctrl-C then re-run in the tmux pane matching *service_name*.
-
-        This is a best-effort approach — it sends C-c followed by Up+Enter
-        to the pane, which re-runs the last command.
-        """
-        # Find the pane by searching for the service name in pane content
-        result = subprocess.run(
-            [
-                "tmux", "list-panes", "-s", "-t", session_name,
-                "-F", "#{pane_id} #{pane_current_command}",
-            ],
-            capture_output=True, text=True,
-        )
-
-        # Fallback: try all panes with C-c + Up + Enter
-        panes = result.stdout.strip().splitlines()
-        for pane_line in panes:
-            pane_id = pane_line.split()[0]
-            # Check if this pane's captured text contains the service name
-            capture = subprocess.run(
-                ["tmux", "capture-pane", "-t", pane_id, "-p"],
-                capture_output=True, text=True,
-            )
-            if service_name in capture.stdout:
-                subprocess.run(["tmux", "send-keys", "-t", pane_id, "C-c", ""])
-                subprocess.run(
-                    ["tmux", "send-keys", "-t", pane_id, "Up", "C-m"],
-                )
-                return
-
-        print(f"  ⚠ Could not find pane for {service_name}")
