+++ local-development/devtool/services/shell_utils.py
@@ -1,0 +1,15 @@
+"""Shared shell helpers for application services and adapters."""
+
+from __future__ import annotations
+
+import shutil
+
+
+def resolve_bash() -> str:
+    """Find bash on PATH instead of assuming /bin/bash."""
+    path = shutil.which("bash")
+    if not path:
+        raise RuntimeError(
+            "bash not found on PATH. Install bash or set SHELL to a compatible shell."
+        )
+    return path
