+++ local-development/services.yaml
@@ -6,24 +6,24 @@ python:
 
 infrastructure:
   mongo:
-    image: "mongo:latest"
+    image: "docker.io/library/mongo:latest"
     ports: ["27017:27017"]
     label: "MongoDB"
     stop_timeout: 30
   rabbitmq:
-    image: "rabbitmq:3-management"
+    image: "docker.io/library/rabbitmq:3-management"
     ports: ["5672:5672", "15672:15672"]
     label: "RabbitMQ"
   qdrant:
-    image: "qdrant/qdrant:latest"
+    image: "docker.io/qdrant/qdrant:latest"
     ports: ["6333:6333", "6334:6334"]
     label: "Qdrant"
   redis:
-    image: "redis:latest"
+    image: "docker.io/library/redis:latest"
     ports: ["6379:6379"]
     label: "Redis"
   temporal:
-    image: "temporalio/temporal:latest"
+    image: "docker.io/temporalio/temporal:latest"
     ports: ["7233:7233", "8233:8233"]
     label: "Temporal"
     command: "server start-dev --ip 0.0.0.0 --ui-port 8233"
